distribution view is now rendered in a QGraphicsView this has the advantage that we can overlay it with partially-transparent widgets in fact, the GUI controls of multi_img_viewer are now placed on top of the display, and the corresponding widget (ViewportControl) fades in/out.
Viewport [QGraphicsScene] -> QGraphicsView (multi_img_viewer.ui)
QGraphicsView -> QGLWidget
depending on what we want to do, we have to work on the scene, on the
graphicsview or on the glwidget. it is quite complicated.

as a next step we need a cleanup of the now 3(!) highly interdependent
custom gui classes (multi_img_viewer, viewport, viewportcontroller)
I will do this as a starting point for separation into model.

another change:
always use QVector::at() where only a const reference is needed.
"at() can be faster than operator[](), because it never causes a deep copy to occur."