quickfix for distribution view subscription crashes
it seems that the check for image.get() in the model was not enough
to avoid harm

this quickfix stores whether a representation's model is currently
subscribed to its data (the view is active) or not

the real solution, as was previously planned, is to delete the payload
(and its model, controller) when it is deactivated (folded-in) by the
user.
however, on reactivation it would need to regain image, label, global
state data (e.g. ignoreLabels)